chore(client): renamed client to AppleMusicAPI
