+++ src/AppleMusicAPI.ts
